feat(powerscale): Extend shared-library events helper with suppression methods
