+++ plugins/module_utils/storage/dell/shared_library/events.py
@@ -208,3 +208,41 @@ class Events:
                 error_message = f'Fetching alert channel failed with error: {error_msg}'
                 LOG.error(error_message)
                 self.module.fail_json(msg=error_message)
+
+    def get_event_suppressed_state(self, event_id):
+        """
+        Get the current suppression state for a specific event
+        :param event_id: Event ID
+        :return: Suppression state (boolean)
+        :rtype: bool
+        """
+        try:
+            result = (self.event_api.get_event_suppress_by_id(
+                event_suppress_id=event_id)).to_dict()
+            return result.get('suppressed', False)
+        except Exception as e:
+            error_msg = utils.determine_error(error_obj=e)
+            error_message = f'Fetching suppression state for event {event_id} failed with error: {error_msg}'
+            LOG.error(error_message)
+            self.module.fail_json(msg=error_message)
+
+    def set_event_suppressed_state(self, event_id, suppressed):
+        """
+        Set the suppression state for a specific event
+        :param event_id: Event ID
+        :param suppressed: Desired suppression state (True/False)
+        :return: Success status
+        :rtype: bool
+        """
+        try:
+            from isilon_sdk.v9_10_0 import EventSuppressIdParams
+            params = EventSuppressIdParams(suppressed=suppressed)
+            (self.event_api.update_event_suppress_by_id(
+                event_suppress_id=event_id,
+                event_suppress_id_params=params))
+            return True
+        except Exception as e:
+            error_msg = utils.determine_error(error_obj=e)
+            error_message = f'Updating suppression state for event {event_id} failed with error: {error_msg}'
+            LOG.error(error_message)
+            self.module.fail_json(msg=error_message)
